LayoutTests/platform/ios-sim/fast/replaced/iframe-with-percentage-height-within-table-with-table-cell-ignore-height-expected.txt

This test checks that iframe with percentage height within table cell ignores the table cell height attribute (strict mode).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL getHeight('iframe-100') should be 460. Was 600.
PASS successfullyParsed is true

TEST COMPLETE
